Moved the mount outside megaraid recipe
There are cases where we may pxe a node and the OS is re-installed.
However, since the node still exists in chef, it will not re-configure
the raid due to the existence of `node['raid']['configured']`.

We want to mount the raid and schedule a reboot, so that it comes back correctly.